refactor: reorganize project structure and update imports
