+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
-import './App.css'
+import './assets/styles/App.css';
 
-function App() {
+export const App = () => {
   const [count, setCount] = useState(0)
 
   return (
@@ -31,5 +31,3 @@ function App() {
     </>
   )
 }
-
-export default App
